Fetch and join data in separate functions
Firstly, methods for a c3 type must respect the method names already
known to the inherited class. We change the names for clarity. Secondly,
for simplicity, we divide the task of JoinData between two methods.

This commit is still probably not quite what the type is supposed to
look like for the purpose of an MLTrainingJob, but I haven't seen a
specific example of how exactly an MLTrainingJob ranges over a list of
latitudes or longitudes or times and instantiates an MLDataSourceSpec
for each combination of values. So this is just another trial for
experimentation purposes.